% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.imf.R
\name{as.mimf}
\alias{as.mimf}
\alias{is.imf}
\title{MIMFs}
\usage{
as.mimf(x, complete = T)

is.imf(x)
}
\arguments{
\item{x}{A vector for \code{is.imf} or an array for \code{as.mimf}.}

\item{complete}{Logical. If TRUE computes the sum of all IMFs in \code{x}
as the recomposed original serie.}
}
\value{
\code{is.imf} returns a logical value indicating if the \code{x} is an
   IMF. 

   \code{as.mimf} returns a \code{mimf} object.
}
\description{
\code{is.imf} tests if its argument is an IMF.
}
\details{
\code{as.mimf} turns an array into a \code{mimf} object.


   \code{is.imf} checks if the vector \code{x} respects the mathematical
   definition of an IMF, i.e. if its difference between the number of extrema
   zero-crossing is at most 1. In addition, computes the relative amplitude
   of the mean enveloppe.

   \code{as.mimf} takes an array and transofrm it in a R object \code{mimf}
   regardless of its mathematical properties. It is made by adding the 
   necessary attributes sch as \code{tt}. It is mostly a convenient function.
}
