% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emdr1.R
\name{choose_lag}
\alias{choose_lag}
\title{Lag IMFs}
\usage{
choose_lag(x, y, lag.max = 0.5)
}
\arguments{
\item{x}{Array containing the IMFs to lag.}

\item{y}{vector containing the response with which to maximize the
correlation.}

\item{lag.max}{The maximum lag to consider relatively to the mean period of
each IMF.}
}
\value{
An array with the same dimensions as \code{x} containing the lagged
   IMFs. In addition contains an attribute \code{lag} storing the lags chosen
   for each IMF.
}
\description{
This function lags the all series in \code{x} according to their cross-
   correlation with \code{y}.
}
\details{
In this function, the lag is determined automatically for each IMF 
   in \code{x}. The lag chosen is the one maximizing the cross-correlation
   (in absolute value) between the IMF and \code{y}.

   For each IMF in \code{x}, the maximum lag considered in the search 
   is lag.max * IMF period (rounded). For example, for an IMF of mean period
   10 and \code{lag.max = 0.5}, the maximum lag considered is 5.
}
\examples{
   choose_lag(cos(3:22), cos(1:20))
   choose_lag(cos(2:21), cos(1:20))

}
