% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mimf_description.R
\name{combine.mimf}
\alias{combine.mimf}
\title{Adds IMFs}
\usage{
combine.mimf(object, select, new.names = NULL)
}
\arguments{
\item{object}{\code{mimf} object or array.}

\item{select}{List of vector giving the IMFs to be summed. Each vector gives
the IMFs to be summed into a single IMF. Elements of \code{select} should
not overlap.}

\item{new.names}{Character vector giving names ti the new IMFS. If NULL
default names are generated.}
}
\value{
Returns a \code{mimf} object.
}
\description{
Sums one or several IMFs and returns a \code{mimf} object. Useful for
   when several IMFs have similar frequencies.
}
\details{
When noise variables are added to the signal to perform the NA-MEMD,
   it often results in a too large number of IMFs with very similar 
   oscillation periods. \code{combine.mimf} allows gathering htese IMFs into 
   a single one. The IMFs to comnine should be carefully selected by 
   checking their mean period and examining the plot.
}
\examples{
   library(dlnm)
   
   # Decompose both temperature and relative humidity with NA-MEMD
   # Adding two noise variables 
   X <- chicagoNMMAPS[,c("temp", "rhum")]
   set.seed(3)
   mimfs <- memd(X, l = 2) # Takes a couple of minutes

   # Plot resulting IMFs
   plot(mimfs)

   # Sum MIMFs with similar frequencies
   cmimfs <- combine.mimf(mimfs, list(11:12, 13:19, 20:21), 
     new.names = c("C11", "C12", "r"))
   plot(cmimfs)

}
