% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emdr2.R
\name{emdr2}
\alias{emdr2}
\title{EMD-R2}
\usage{
emdr2(formula, mimf, covariates = NULL, tt = attr(mimf, "tt"),
  reg.fun = "glm", reg.args = list(), pimf.args = list(),
  xy.args = c("x", "y"))
}
\arguments{
\item{formula}{A formula describing the models to be fit for each MIMF.}

\item{mimf}{\code{mimf} object containing both the response and predictor
IMFs.}

\item{covariates}{A list or data.frame containing eventual covariates
which are not IMFs.}

\item{tt}{A vector of time indices for \code{mimf}. The default is to use
the \code{tt} attribute.}

\item{reg.fun}{The regression function to perform either as a function or
as a character. Should include \code{formula} and \code{data} arguments.}

\item{reg.args}{A list containing additional arguments to be passed to 
\code{reg.fun}.}

\item{pimf.args}{A list of arguments to be passed to the function 
\code{\link{pimf}} to prepare the IMFs for the regression.}

\item{xy.args}{A character vector of length two indicating the argument
names corresponding to the predictors and response in the case 
\code{reg.fun} does not include arguments \code{formula} and \code{data}.
The default of \code{xy.args = c("x", "y")} should cover the majority
of cases.}
}
\value{
A \code{emdr2} object which is a list of length nimf containing all
   results of \code{reg.fun} on individual MIMFs.
}
\description{
Performs EMD-R2 by using the given regression function in each MIMF.
}
\details{
\code{emdr2} wraps the chosen regression function (\code{reg.fun}),
   and applies it once per MIMF in \code{mimf}. Variables given in
   \code{covariates} are used with each IMF.

   \code{reg.fun} can include either a couple \code{formula} and \code{data}
   arguments, either \code{x} and \code{y} arguments. Particular
   argument names for the response or predictor variables can be given
   in \code{xy.args}.
}
\examples{
   library(dlnm)
   library(glmnet)

   dat <- chicagoNMMAPS[,c("death", "temp", "rhum")]

   mimfs <- memd(dat)
   cmimfs <- combine.mimf(mimfs, list(12:13, 14:17, 18:19), 
     new.names = c("C12", "C13", "r"))
   amps <- mean_amplitude(cmimfs)

   # EMD-R2 with glm
   lm.R2 <- emdr2(death ~ temp + rhum, mimf = cmimfs)
   betas.R2 <- coef(lm.R2)
   sensitivity.R2 <- sensitivity(amps[,-1], coefs = betas.R2[,-1])
   plot_emdr(sensitivity.R2, periods = period(cmimfs)[,-1],  
     col = c("red", "blue"), pch = 16:17)
   abline(h = 0, lty = 2)
   
   # EMD-R2 with lasso
   lasso.R2 <- emdr2(death ~ temp + rhum, mimf = cmimfs, 
     reg.fun = "cv.glmnet")
   betas.R2 <- coef(lasso.R2, s = "lambda.1se")
   sensitivity.R2 <- sensitivity(amps[,-1], coefs = betas.R2[,-1])
   plot_emdr(sensitivity.R2, periods = period(cmimfs)[,-1], 
     show.coef = "nonzero", pch = 16:17)
   abline(h = 0, lty = 2)

}
\references{
Masselot, P., Chebana, F., Belanger, D., St-Hilaire, A., Abdous, B., 
   Gosselin, P., Ouarda, T.B.M.J., 2018. EMD-regression for modelling 
   multi-scale relationships, and application to weather-related 
   cardiovascular mortality. \emph{Science of The Total Environment} 
   612, 1018-1029.
}
\seealso{
\code{\link{extract.emdr2}} to extract a particular element of the
   result from \code{reg.fun} for each submodel. \code{\link{coef.emdr2}}
   to specifically extract the coefficients. \code{\link{predict.emdr2}}
   to obtain predictions from the whole EMD-R2 model.
}
