% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imf_test.R
\name{imf.test}
\alias{imf.test}
\alias{print.imftest}
\title{IMF significance test}
\usage{
imf.test(object, tt = NULL, type = c("wu", "flandrin"),
  alternative = c("one.sided", "two.sided"), H = 0.5)

\method{print}{imftest}(x, digits = 3, ...)
}
\arguments{
\item{object}{\code{mimf} object or array containing (potentially
multivariate) IMFs.}

\item{tt}{Numeric vector of custom time indices.}

\item{type}{The type of significance test to perform. Either the test of 
Wu and Huang (2004) (\code{type = "wu"}) or the test of Flandrin et al. 
(2004) (\code{type = "flandrin"}).}

\item{alternative}{Character giving if the alternative hypothesis is
one-sided (\code{alternative = "one.sided"}) or two-sided
(\code{alternative = "two.sided"}).}

\item{H}{Numeric value giving the Hurst exponent necessary for the test of
Flandrin et al. (2004). Must be between 0 and 1.}

\item{x}{An \code{imftest} object resulting from a call to \code{imf.test}.}

\item{digits}{The number of digits to print p-values.}

\item{...}{Formatting arguments to be passed to the function 
\code{\link[base]{format}}}
}
\value{
An object of class \code{imftest} containing the following elements:
     \item{object}{The input \code{mimf} object name.}
     \item{alternative}{The alternative hypothesis.}
     \item{type}{The type of test performed.}
     \item{logp}{Matrix containing the log mean-period of each IMF.}
     \item{loge}{Matrix containing the log energy (variance) of each IMF.}
     \item{p.value}{Matrix containing the p-value of the test for each IMF.}
}
\description{
Tests the hypothesis that IMFs do not contain more information than white
   noise.
}
\details{
In both tests, the first (highest frequency) IMF is used to estimate the
   noise level of the input signal and is thus never significant.
}
\examples{
   library(dlnm)
   
   # Decompose both temperature and relative humidity with NA-MEMD
   # Adding two noise variables 
   X <- chicagoNMMAPS[,c("temp", "rhum")]
   set.seed(3)
   mimfs <- memd(X, l = 2) # Takes a couple of minutes
   cmimfs <- combine.mimf(mimfs, list(11:12, 13:19, 20:21), 
     new.names = c("C11", "C12", "r"))
   
   # Apply the test of Wu and Huang (2004)
   testres <- imf.test(cmimfs)
   testres
   plot(testres)

}
\references{
Wu, Z., Huang, N.E., 2004. A study of the characteristics of white noise 
   using the empirical mode decomposition method. \emph{Proceedings of the 
   Royal Society of London. Series A: Mathematical, Physical and Engineering 
   Sciences} 460, 1597-1611.

   Flandrin, P., Goncalves, P., Rilling, G.G., 2004. Detrending and 
   denoising with empirical mode decompositions. Presented at the \emph{12th 
   European Signal Processing Conference}, IEEE, Vienna, Austria.
}
