% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emdr1.R
\name{pimf}
\alias{pimf}
\title{EMD-R1 preparation}
\usage{
pimf(x, y, covariates = NULL, tt = attr(x, "tt"), lag = NA,
  lag.max = 0.5, fill.na = FALSE)
}
\arguments{
\item{x}{\code{mimf} object to be prepared for regression.}

\item{y}{Vector containing the response variable of the regression.}

\item{covariates}{List containing eventual covariates to be added to the 
output data.frame.}

\item{tt}{Vector of custom time indices. Useful to lag irregular time series.}

\item{lag}{Vector of lagging values for the IMFs in \code{x}. If NA (the
default) lags are automatically chosen by the function 
\code{\link{choose_lag}}.}

\item{lag.max}{The maximum lag to consider in the automatic lag choice. See
\code{\link{choose_lag}}.}

\item{fill.na}{Logical indicating if irregularly sampled IMFs should be
   filled with NAs.

This function take a \code{mimf} object, a response vector and prepare them
   to be easily used in any regression function. It essentially transform
   the MIMF array into a matrix containing all IMFs and lags these
   IMFs if necessary.

Also allows the inclusion of non-IMF predictors in \code{covariates} that
   are included in the output data.frame.}
}
\value{
A data.frame containing all necessary variables for a regression
   function. The first column contains the response vector \code{y}. The
   following columns contain the IMFs to be used and the last columns 
   contain the eventual covariates.
}
\description{
Prepare a data.frame to be used in a regression function for EMD-R1, 
   including the predictors MIMF and the response.
}
\examples{
   library(dlnm)
   
   # Predictor decomposition
   X <- chicagoNMMAPS[,c("temp", "rhum")]
   set.seed(3)
   mimfs <- memd(X, l = 2) # Takes a couple of minutes
   cmimfs <- combine.mimf(mimfs, list(11:12, 13:19, 20:21), 
     new.names = c("C11", "C12", "r"))

   # Response variable
   Y <- chicagoNMMAPS$resp[attr(cmimfs, "tt")]

   # Data preparation: includes the day-of-week variable as potential
   # confounder
   dataR1 <- pimf(cmimfs, Y, covariates = list(dow = 
     chicagoNMMAPS$dow[attr(cmimfs, "tt")]))
   
   # Apply the Lasso
   library(glmnet)
   lasso.res <- cv.glmnet(data.matrix(dataR1[,-1]), dataR1[,1], 
     family = "poisson")

   # Compute sensitivity and plot results
   amps <- mean_amplitude(dataR1[,2:27])
   betas <- coef(lasso.res)
   s <- sensitivity(amps, coefs = betas[2:27]) 
   plot_emdr(matrix(s, ncol = 2, byrow = FALSE), periods = period(cmimfs), 
   show.coef = "nonzero", col = c("red", "blue"), pch = 16:17)
   abline(h = 0, lty = 2)

}
\references{
Masselot, P., Chebana, F., Belanger, D., St-Hilaire, A., Abdous, B., 
   Gosselin, P., Ouarda, T.B.M.J., 2018. EMD-regression for modelling 
   multi-scale relationships, and application to weather-related 
   cardiovascular mortality. \emph{Science of The Total Environment} 
   612, 1018-1029.
}
