% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imf_test.R
\name{plot.imftest}
\alias{plot.imftest}
\title{Graphical IMF significance test}
\usage{
\method{plot}{imftest}(
  x,
  level = 0.05,
  select.var = NULL,
  select.imf = NULL,
  alternative = x$alternative,
  nline = 500,
  col = "black",
  col.signif = NULL,
  pch = 4,
  pch.signif = 8,
  imf.names = NULL,
  col.names = NULL,
  col.lines = "black",
  lty.lines = 2,
  lwd.lines = 1,
  add = FALSE,
  ...
)
}
\arguments{
\item{x}{An \code{imftest} x resulting from a call to 
\code{link{imf.test}}.}

\item{level}{The significance level at which to draw the rejectance region
lines.}

\item{select.var}{Character of numeric vector giving a subset of variables 
to plot.}

\item{select.imf}{Character of numeric vector giving a subset of IMFs 
to plot.}

\item{alternative}{The alternative hypothesis, either "one.sided" or 
"two.sided".}

\item{nline}{Numeric value giving the resolution of the rejectance region 
lines.}

\item{col}{The colors of the points.}

\item{col.signif}{The color to use for significant IMFs exclusively. If NULL,
\code{col} is used.}

\item{pch}{The plotting character of IMFs.}

\item{pch.signif}{Specific \code{pch} for significant IMFs only.}

\item{imf.names}{Character vector of IMF labels to plot next to the points.}

\item{col.names}{Color of the IMF labels.}

\item{col.lines}{Colors of the significance lines.}

\item{lty.lines}{Significance lines type.}

\item{lwd.lines}{Significance lines width.}

\item{add}{Logical value. If TRUE, the plot is added to an existing one.}

\item{...}{Additional graphical parameters. See \code{\link[graphics]{par}}.}
}
\description{
Graphically displays the result of an IMF significance test by plotting
   the log energy vs. log period of IMFs.
}
\examples{
   library(dlnm)
   
   # Decompose both temperature and relative humidity with NA-MEMD
   # Adding two noise variables 
   X <- chicagoNMMAPS[,c("temp", "rhum")]
   set.seed(123)
   mimfs <- memd(X, l = 2) # Takes a couple of minutes
   cmimfs <- combine.mimf(mimfs, list(10:11, 12:13), 
     new.names = c("C10", "C11"))
   
   # Apply the test of Wu and Huang (2004)
   testres <- imf.test(cmimfs)
   plot(testres)

}
\references{
Wu, Z., Huang, N.E., 2004. A study of the characteristics of white noise 
   using the empirical mode decomposition method. \emph{Proceedings of the 
   Royal Society of London. Series A: Mathematical, Physical and Engineering 
   Sciences} 460, 1597–1611.
}
