% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mimf_description.R
\name{plot.mimf}
\alias{plot.mimf}
\title{Plot IMFs}
\usage{
\method{plot}{mimf}(
  x,
  tt = NULL,
  select.var = NULL,
  select.imf = NULL,
  input = TRUE,
  input.lab = "X",
  imf.lab = NULL,
  grid = c("zeroline", "complete", "none"),
  grid.col = "lightgray",
  space = 1,
  add.legend = p > 1,
  legend.pars = list(),
  ...
)
}
\arguments{
\item{x}{\code{mimf} x or array to plot.}

\item{tt}{Vector containing custom time indices for the IMFs. If NULL, looks
for the \code{tt} attribute of \code{x}.}

\item{select.var}{Character or numeric vector giving a subset of variables 
for which to plot the IMFs.}

\item{select.imf}{Character or numeric vector giving a subset of IMFs to 
plot.}

\item{input}{Logical. If TRUE, the top panel shows the original signal.
Only considered if \code{x} is a \code{mimf} object.}

\item{input.lab}{The label of the panel showing the input signal.}

\item{imf.lab}{Character vector giving labels for the IMFs. NULL displays the
dimnames of \code{x} and NA removes the labels.}

\item{grid}{Character giving the type of grid to plot. "zeroline" 
(the default) only draws the zeroline, "complete" draws a complete grid 
and "none" draws no grid at all.}

\item{grid.col}{The grid color.}

\item{space}{Numeric value giving the margin between two panels in the plot.}

\item{add.legend}{Logical value. If TRUE (the default) a legend is
automatically drawn at the top of the plot.}

\item{legend.pars}{List of optional parameters for the legend.
Can be useful to indicate custom names for the variables for instance.
See \code{\link[graphics]{legend}}.}

\item{...}{Other graphical parameters. See 
\code{\link[graphics]{par}}.}
}
\description{
Method to display the (M)IMFs obtained by the function \code{\link{memd}}.
}
\details{
One panel is drawn for each IMF. In the multivariate case, 
   by default all IMF's variables are scaled and displayed on the same panel.
   To obtain the true amplitude of each variable, they must be plotted
   separately.

   If noise channel are present in the signal, i.e. if the argument
   \code{keep.noise = TRUE} in \code{\link[emdr]{memd}},
   they are not displayed by default. To display them, the argument 
   \code{select.var} must be set manually. Note that, in this case, it
   automatically set the argument \code{input} to \code{FALSE}.
}
\examples{
   library(dlnm)
   
   # Decompose both temperature and relative humidity with NA-MEMD
   # Adding two noise variables 
   X <- chicagoNMMAPS[,c("temp", "rhum")]
   mimfs <- memd(X, l = 2) # Takes a couple of minutes

   # Plot the two variables on the same graph
   plot(mimfs)

   # Plot the two variables separately
   plot(mimfs, select.var = "temp", col = "red")
   plot(mimfs, select.var = "rhum", col = "blue")

}
