% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emdr_results.R
\name{plot_emdr}
\alias{plot_emdr}
\title{Plot coefficients of EMD-regression}
\usage{
plot_emdr(x, periods = NULL, lower = NULL, upper = NULL,
  ci.args = list(), period.log2 = TRUE, trend.label = "Trend",
  show.coef = c("all", "nonzero", "significant"), col = NULL,
  pch = NULL, line.pars = list(), ...)
}
\arguments{
\item{x}{The coefficient matrix to plot. Can also contain sensitivities
(see \code{\link{sensitivity}}).}

\item{periods}{Matrix containting the mean period of IMFs correspondind to
the coefficients in \code{x}. See \code{\link{period}}. If NULL,
the period are taken as the two to the power of the IMF's order.}

\item{lower, upper}{Matrices containing lower and upper confidence limits.}

\item{ci.args}{A list of arguments to be passed to the function
\code{\link[graphics]{arrows}} for drawing confidence intervals.}

\item{period.log2}{Logical. If TRUE, a log2 transformation is applied to the
x axis.}

\item{trend.label}{The label to be displayed for the trend
component's coefficient.}

\item{show.coef}{Character giving restrictions for the coefficients to draw.
\code{show.coef = "all"} (the default) draws all coefficients, 
\code{show.coef = "nonzero"} plots only nonzero coefficients, 
\code{show.coef = "significant"} draws only coefficients for which the
confidence interval given in \code{lower} and \code{lower} excludes 
the zero value.}

\item{col, pch}{color and point type for drawn coefficients. If matrices one
value corresponds to one coefficient and if vectors a value per variable
 is used.}

\item{line.pars}{List of parameters for the line separating the trend
coefficients from the other. See \code{\link[graphics]{abline}}.}

\item{...}{Other arguments to be passed to the plot. See 
\code{\link[graphics]{par}}.}
}
\description{
Plot the coefficients resulting from an EMD-regression acording to the
   mean period of the corresponding IMFs.
}
\examples{
   ## EMD-R1
   library(dlnm)
   library(glmnet)
   
   # Predictor decomposition
   X <- chicagoNMMAPS[,c("temp", "rhum")]
   set.seed(3)
   mimfs <- memd(X, l = 2) # Takes a couple of minutes
   cmimfs <- combine.mimf(mimfs, list(11:12, 13:19, 20:21), 
     new.names = c("C11", "C12", "r"))

   # Response variable
   Y <- chicagoNMMAPS$resp[attr(cmimfs, "tt")]

   # Data preparation: includes the day-of-week variable as potential
   # confounder
   dataR1 <- pimf(cmimfs, Y, covariates = list(dow = 
     chicagoNMMAPS$dow[attr(cmimfs, "tt")]))
   
   # Apply the Lasso
   library(glmnet)
   lasso.res <- cv.glmnet(data.matrix(dataR1[,-1]), dataR1[,1], 
     family = "poisson")

   # Compute sensitivity and plot results
   amps <- mean_amplitude(dataR1[,2:27])
   betas <- coef(lasso.res)
   s <- sensitivity(amps, coefs = betas[2:27]) 

   plot_emdr(matrix(s, ncol = 2, byrow = FALSE), periods = period(cmimfs), 
   show.coef = "nonzero", col = c("red", "blue"), pch = 16:17)
   abline(h = 0, lty = 2) 

   ## EMD-R2
   dat <- chicagoNMMAPS[,c("death", "temp", "rhum")]

   set.seed(3475)
   mimfs <- memd(dat, l = 2, wn.power = .1)
   cmimfs <- combine.mimf(mimfs, list(12:14, 15:19, 20:22), 
     new.names = c("C12", "C13", "r"))

   # EMD-R2 with glm
   lm.R2 <- emdr2(death ~ temp + rhum, mimf = cmimfs)
   betas.R2 <- coef(lm.R2)
   amps <- mean_amplitude(cmimfs)
   sensitivity.R2 <- sensitivity(amps[,-1], coefs = betas.R2[,-1])

   plot_emdr(sensitivity.R2, periods = period(cmimfs)[,-1],  
     col = c("red", "blue"), pch = 16:17)
   abline(h = 0, lty = 2)

}
