% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emdr_results.R
\name{sensitivity}
\alias{sensitivity}
\title{Response sensitivity to IMFs}
\usage{
sensitivity(amplitudes, model = NULL, coefs = NULL, ...)
}
\arguments{
\item{amplitudes}{A matrix of mean amplitudes as computed by 
\code{\link{mean_amplitude}}. One column corresponds to a variable and
one line to an IMF.}

\item{model}{The result from a regression function. Must have a 
\code{\link[stats]{coef}} method. If not, use the argument \code{coefs}
instead.}

\item{coefs}{User provided coefficient matrix. Must have the same dimensions
as \code{amplitudes}.}

\item{...}{Additional arguments for the \code{\link[stats]{coef}} method
used by \code{model}.}
}
\value{
A matrix of sensitivities, with the same dimensions as 
   \code{amplitudes}.
}
\description{
This function computes the sensitivity to each IMF, which is basically a
   regression coefficient scaled by the IMF mean amplitude.
}
\details{
The sensitivy term hereby designates regression coefficients
   scaled according to the corresponding IMF's mean amplitude. It estimates
   the amplitude of response's variations explained by the IMF.

   The function uses the results from a regression model to compute 
   the sensitivity.
   If the resulting object contains a \code{\link[stats]{coef}} method
   it is used to extract the necessary coefficients. If this is not the 
   case, the \code{coefs} argument must be used instead.
}
\examples{
   ## EMD-R1
   library(dlnm)
   library(glmnet)
   
   # Predictor decomposition
   X <- chicagoNMMAPS[,c("temp", "rhum")]
   set.seed(3)
   mimfs <- memd(X, l = 2) # Takes a couple of minutes
   cmimfs <- combine.mimf(mimfs, list(11:12, 13:19, 20:21), 
     new.names = c("C11", "C12", "r"))

   # Response variable
   Y <- chicagoNMMAPS$resp[attr(cmimfs, "tt")]

   # Data preparation: includes the day-of-week variable as potential
   # confounder
   dataR1 <- pimf(cmimfs, Y, covariates = list(dow = 
     chicagoNMMAPS$dow[attr(cmimfs, "tt")]))
   
   # Apply the Lasso
   library(glmnet)
   lasso.res <- cv.glmnet(data.matrix(dataR1[,-1]), dataR1[,1], 
     family = "poisson")

   # Compute sensitivity and plot results
   amps <- mean_amplitude(dataR1[,-1])
   betas <- coef(lasso.res)
   s <- sensitivity(amps, coefs = betas[-1]) 

   ## EMD-R2
   dat <- chicagoNMMAPS[,c("death", "temp", "rhum")]

   set.seed(3475)
   mimfs <- memd(dat, l = 2, wn.power = .1)
   cmimfs <- combine.mimf(mimfs, list(12:14, 15:19, 20:22), 
     new.names = c("C12", "C13", "r"))

   # EMD-R2 with glm
   lm.R2 <- emdr2(death ~ temp + rhum, mimf = cmimfs)
   betas.R2 <- coef(lm.R2)
   amps <- mean_amplitude(cmimfs)
   sensitivity.R2 <- sensitivity(amps[,-1], coefs = betas.R2[,-1])

}
\references{
Masselot, P., Chebana, F., Bélanger, D., St-Hilaire, A., Abdous, B., 
   Gosselin, P., Ouarda, T.B.M.J., 2018. EMD-regression for modelling 
   multi-scale relationships, and application to weather-related 
   cardiovascular mortality. \emph{Science of The Total Environment} 
   612, 1018–1029.
}
\seealso{
\code{link{coef.emdr2}} to extract coefficients from an \code{emdr2}
   object.
}
